Commented unused parameters to suppress warnings in Visual Studio 2005: src/vw/Image/ImageView.h src/vw/Image/ImageResourceView.h src/vw/Camera/CameraModel.h src/vw/Math/Vector.h src/vw/FileIO/DiskImageResourcePDS.h
src/vw/FileIO/DiskImageResource.h: changed unsigned int to int32 to remove error from call to std::min

src/vw/Image/Convolution.h: cast to int before negation of unsigned int

Added new progress callback class to report progress from quadtree generation
src/vw/Core/ProgressCallback.cc
src/vw/Core/ProgressCallback.h
src/vw/Core/Exception.h:  Added new exception type:  Aborted

src/vw/Mosaic/QuadTreeGenerator.h: added ProgressCallback
src/vw/Mosaic/qtree.cc: added ProgressCallback

+++ src/vw/Image/Convolution.h
@@ -383,7 +383,7 @@ namespace vw {
       typedef typename ProductType<typename SrcT::pixel_type, KernelT>::type AccumT;
 
       SrcAccessT splane = src.origin();
-      EdgeAccessT eplane = edge_view.origin().advance(-c,0);
+      EdgeAccessT eplane = edge_view.origin().advance(-(int)c,0);
       DestAccessT dplane = dest.origin();
       for( unsigned p=0; p<src.planes(); ++p ) {
         SrcAccessT srow = splane;
